Trade: extract reusable code out of magnum-imageconverter.
Will be needed by magnum-sceneconverter as well.
